requirements.txt
asgiref==3.8.1
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==5.0.6
django-debug-toolbar==4.4.6
django-prometheus==2.3.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.1
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
packaging==24.1
pillow==10.3.0
platformdirs==4.3.6
pre_commit==4.0.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.6
